requirements.txt
python-telegram-bot>=20.0
gspread>=5.0
google-auth>=2.0
Flask>=2.3.0gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
Flask
gunicorn
gunicorn>=20.1.0gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
Flask
gunicorn
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
Flask
gunicorn